chore!: rename `TGeo` to `Root` plugin (#4352)
This PR renames the `TGeo` plugin to `Root` plugin and adapts the
corresponding flags and documentation.

This is in preparation for moving some 'encapsulated` functionality from
the Examples to the Plugins, such as material track writing, in order to
provide them consistently to client experiments.

Since the code was re-parsed by `SonarCloud` I fixed most of the issues.

--- END COMMIT MESSAGE ---

Any further description goes here, @-mentions are ok here!

- Use a *conventional commits* prefix: [quick
summary](https://www.conventionalcommits.org/en/v1.0.0/#summary)
- We mostly use `feat`, `fix`, `refactor`, `docs`, `chore` and `build`
types.
- A milestone will be assigned by one of the maintainers

+++ Plugins/Root/src/TGeoMaterialConverter.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "Acts/Plugins/TGeo/TGeoMaterialConverter.hpp"
+#include "Acts/Plugins/Root/TGeoMaterialConverter.hpp"
 
 #include "Acts/Material/Material.hpp"
 
@@ -25,9 +25,12 @@ Acts::MaterialSlab Acts::TGeoMaterialConverter::materialSlab(
   double uScalor = options.unitLengthScalor;
   double rScalar = options.unitMassScalor / pow(options.unitLengthScalor, 3);
 
-  auto material = Material::fromMassDensity(
-      matX0 * uScalor / cFactor, matL0 * uScalor / cFactor, tgMaterial.GetA(),
-      tgMaterial.GetZ(), matRho * rScalar * cFactor);
+  auto material =
+      Material::fromMassDensity(static_cast<float>(matX0 * uScalor / cFactor),
+                                static_cast<float>(matL0 * uScalor / cFactor),
+                                static_cast<float>(tgMaterial.GetA()),
+                                static_cast<float>(tgMaterial.GetZ()),
+                                static_cast<float>(matRho * rScalar * cFactor));
 
-  return MaterialSlab(material, thicknessOut);
+  return MaterialSlab(material, static_cast<float>(thicknessOut));
 }
